refactored AppImage Update Information parser.
